requirements.txt
flask
flask_restx
requests
transformers
pillow
openai